+++ JOSS/paper.bib
@@ -28,7 +28,7 @@
   number = {8},
   pages = {2007--2023},
   year = {2011},
-  doi={dx.doi.org/10.1021/ci200217w},
+  doi={10.1021/ci200217w},
 }
 @article{thomas2013computing,
   title={Computing vibrational spectra from ab initio molecular dynamics},
