### 2026-01-19 ddf12ab
reorganize the paper
