+++ JOSS/paper.md
@@ -55,23 +55,35 @@ Vibrational spectroscopy, particularly IR and Raman techniques, provides valuabl
 
 Computational simulation of vibrational spectroscopy is often not straightforward. Most quantum chemistry packages that perform electronic structure calculations provide only the dipole moments or polarizability tensors, rather than the final spectrum itself. The challenge becomes greater when one goes beyond the static regime to compute vibrational spectra, such as when performing MD simulations to capture anharmonic modes or solvent effects, or real-time time-dependent density functional theory (RT-TDDFT) simulations to access excited state dynamics, relevant for RR spectroscopy.
 
-Post-processing programs that can handle these properties are often limited in scope and scattered across different software packages. In this work, we introduce Vibrant, a computational tool which bridges this gap by performing a series of post-processing procedures on dipole moment and polarizability data to generate MD-based or static vibrational spectra. Our code can provide these spectra for gas-phase, liquids and a wide range of materials. Figure&#160;1 provides an overview of Vibrant's major functionalities.
+Post-processing programs that can handle these properties are often limited in scope and scattered across different software packages. In this work, we introduce Vibrant, a computational tool which bridges this gap by performing a series of post-processing procedures on dipole moment and polarizability data to generate MD-based or static vibrational spectra. Our code can provide these spectra for gases, liquids and a wide range of materials. Figure&#160;1 provides an overview of Vibrant's major functionalities.
 
-The first panel (top-left) summarizes Vibrant’s frequency analysis capabilities, which include static calculations within the harmonic approximation and post-processing of MD trajectories. In the static approach, conventional normal mode analysis is used to obtain vibrational frequencies and normal mode coordinates, while in the MD-based approach, normal mode derivatives are replaced by Fourier transforms of time-autocorrelation functions, yielding so-called power spectra [@thomas2013computing; @gordon1965molecular].
+The first panel (top-left) summarizes Vibrant’s frequency analysis capabilities, which include static calculations within the harmonic approximation and post-processing of MD trajectories. In the static approach, conventional normal mode analysis is used to obtain vibrational frequencies and normal mode coordinates, while in the MD-based approach, normal mode derivatives are replaced by Fourier transforms of time-autocorrelation functions, yielding power spectra [@thomas2013computing; @gordon1965molecular].
 
-The second panel (top-right) highlights Vibrant’s functionality for computing IR, non-resonant Raman and RR spectral intensities, which can be obtained from either static or MD-based calculations. For IR and non-resonant Raman spectroscopy, Vibrant can handle different types of dipole moments and polarizabilities. Dipole moments may be provided as Berry-phase dipoles or generated from Wannier centers, while polarizabilities can be taken from density functional perturbation theory (DFPT) or computed from induced dipole moments via a finite-difference approach. RR intensities are obtained by post-processing time-dependent (Berry-phase) dipole trajectories RT-TDDFT simulations. Although not shown in the panel, Vibrant also supports the computation of absorption spectra from RT-TDDFT dipole data. 
[…]
