+++ JOSS/paper.bib
@@ -324,4 +324,26 @@
   year={2020},
   doi = {10.21105/joss.02671} 
 }
-
+@article{feuerstein2025nitrile,
+  title={Nitrile Groups as Build-In Molecular Sensors for Interfacial Effects at Electrocatalytically Active Carbon--Nitrogen Materials},
+  author={Feuerstein, Linda and Bas, Ekin Esme and Golze, Dorothea and Heine, Thomas and Oschatz, Martin and Weidinger, Inez M},
+  journal={ACS applied materials \& interfaces},
+  volume={17},
+  number={16},
+  pages={23996--24004},
+  year={2025},
+  publisher={ACS Publications},
+  doi = {10.1021/acsami.5c02366} 
+}
+@dataset{winogradow2025,
+  author       = {Winogradow, Ekin and
+                  Scheffler, Johannes and
+                  Leucke, Moritz and
+                  Heine, Thomas and
+                  Golze, Dorothea},
+  title        = {Example calculations for Vibrant},
+  month        = nov,
+  year         = 2025,
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.17573395}
+}
