+++ JOSS/paper.md
@@ -77,13 +77,13 @@ To the best of our knowledge, no existing computational tool provides the same l
 
 # Research impact statement
 
-Over the course of its development, we used Vibrant in several publications to calculate and analyze vibrational spectra. The most prominent example is its application to the MD-based and static IR and Raman spectra of the layered framework material COF-1, processing different types of dipole moments and polarizabilities [@bas2024robust]. In the same study, we also used Vibrant to investigate solvent contributions by decomposing the MD-based spectra into contributions from the framework building blocks and solvent molecules, thereby facilitating spectral analysis. In addition, we used Vibrant to generate static Raman spectra for nitrile-containing molecular C~2~N frameworks [@feuerstein2025nitrile] and to compute Padé-interpolated RT-TDDFT-based absorption spectra for gas-phase naphthalene as part of the GreenX library documentation [@leucke2025analytic]. In all these studies, the results were compared against other available computational tools when applicable and showed excellent agreement with them, and also with experiment.
+Over the course of its development, we used Vibrant in several publications to calculate and analyze vibrational spectra. The most prominent example is its application to the MD-based and static IR and Raman spectra of the layered framework material COF-1, processing different types of dipole moments and polarizabilities [@bas2024robust]. In the same study, we also used Vibrant to investigate solvent contributions by decomposing the MD-based spectra into contributions from the framework building blocks and solvent molecules, thereby facilitating spectral analysis. In addition, we used Vibrant to generate static Raman spectra for nitrile-containing molecular C~2~N frameworks [@feuerstein2025nitrile] and to compute Padé-interpolated RT-TDDFT-based absorption spectra for gas-phase naphthalene as part of the GreenX library documentation [@leucke2025analytic]. In all these studies, the results were compared against other available computational tools when applicable and showed excellent agreement with them and with experiment.
 
-Apart from its documented applications, Vibrant provides extensive documentation covering its functionalities, implementation details, and underlying theory on the [Vibrant website](https://lstsgroup.github.io/vibrant/index.html) Provided tutorials and input/output repositories [@winogradow2025] enable user-friendly testing, while its open-source nature and regression testing promotes possible development and extension of its features.
+Apart from its documented applications, Vibrant provides extensive documentation covering its functionalities, implementation details, and underlying theory on [its website](https://lstsgroup.github.io/vibrant/index.html). Provided tutorials and repositories containing input/output files of test calculations [@winogradow2025] enable user-friendly testing, while the open-source nature of the code and its regression testing support future development and extension of its features.
 
 # AI usage disclosure
 
-No generative AI tools were used for developing the Vibrant software package and preparing its documentation, authoring this manuscript, or producing the supplementary materials.
+No generative AI tools were used for developing the Vibrant package and preparing its documentation, authoring this manuscript, or producing the supplementary materials.
 
 # Acknowledgements
 
