+++ JOSS/paper.md
@@ -73,7 +73,7 @@ Vibrant is an open-source post-processing tool for vibrational spectroscopy dist
 
 A limited number of computational programs are available for calculating MD-based and static vibrational spectra. A prominent example is TRAVIS [@brehm2020travis; @brehm2011travis], which processes various properties obtained from MD simulations to compute different types of vibrational spectra, although it is mainly designed for liquids. There are also Python-based tools for vibrational analysis, including the scripts that process CP2K [@kuhne2020cp2k] polarizabilities [@mattiat2018efficient] to generate static Raman spectra [@beat_hubmann_2020_4026342], the scripts distributed with FHI-aims [@blum2009ab; @abbott2025roadmap] for normal mode analysis and static IR and Raman calculations, and VibIR-Parallel-Compute, which focuses on efficient static IR spectra [@rojas2025vibir]. For materials, a popular tool is Phonopy [@phonopy-phono3py-JPCM; @phonopy-phono3py-JPSJ], which calculates phonon dispersion relations and vibrational densities of state; however, it is not designed for MD-based spectral analysis. Related tools include FHI-vibes [@knoop2020fhi], which integrates Phonopy to automate vibrational and MD workflows for FHI-aims, and THeSeuSS [@boziki2025journey], which computes static IR and Raman spectra interfacing with FHI-aims.
 
-To the best of our knowledge, no existing computational tool provides the same level of flexibility as Vibrant, which supports a wide range of vibrational spectra, both static and MD-based methodologies, diverse dipole and polarizability types, and applicability across gaseous, liquid, and solid-state systems.
+To the best of our knowledge, no existing computational tool provides the same level of flexibility as Vibrant, which supports a wide range of vibrational spectra, both static and MD-based methodologies, diverse dipole and polarizability types, and applicability across gaseous, liquid, and solid-state systems. In this context, extending existing tools such as TRAVIS was not feasible due to their limited support for community-driven development, while other mentioned packages are either much smaller in scope or not designed for broad extensibility. The need to interface with external libraries such as GreenX for computationally demanding tasks further motivated the development of Vibrant as an independent package.
 
 # Research impact statement
 
