### 2026-01-21 #9594 @ekinesme
Dear @arfon

Thank you for the information regarding the new guidelines of JOSS. We updated our paper according to the new regulations.

Best!
