### 2026-01-28 fa1e0ba
edit paper.md
