### 2026-03-02 #9594 @espottesmith
Hey @ekinesme,

I'm Evan Spotte-Smith, and I'll be serving as the Topic Editor for this submission. To get the ball rolling, could you suggest some potential reviewers? You could look at similar JOSS submissions or our [reviewer database](https://reviewers.joss.theoj.org/reviewers) for inspiration.
