### 2026-03-02 #9594 @moritzleucke
> Hey [@ekinesme](https://github.com/ekinesme),
>
> I'm Evan Spotte-Smith, and I'll be serving as the Topic Editor for this submission. To get the ball rolling, could you suggest some potential reviewers? You could look at similar JOSS submissions or our [reviewer database](https://reviewers.joss.theoj.org/reviewers) for inspiration.

Hey @espottesmith,

Thanks for editing this submission!
Here is our proposal for potential reviewers (posted earlier in this thread): https://github.com/openjournals/joss-reviews/issues/9594#issuecomment-3641701339
