### 2026-03-02 #9594 @espottesmith
@raghurama123 @mbagheri20 @mailhexu @ksunden

Can any of y'all contribute your expertise to this submission to JOSS?
