### 2026-03-02 #10136 @editorialbot
[REVIEW]: Vibrant: A Post-Processing Tool for Computational Vibrational Spectroscopy of Molecules, Liquids and Materials
**Submitting author:** @ekinesme (<a href="http://orcid.org/0000-0002-0110-4691">Ekin Esme Winogradow</a>)
**Repository:** https://github.com/lstsgroup/vibrant
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
