### 2026-03-24 #10136 @mailhexu
@espottesmith I’ve reviewed the paper and documentation, and have also tested the code. This is a really valuable contribution to the community. The paper is very well-written, and the documentation is clear and easy to follow. Both the installation and the test runs went smoothly. I recommend the publication.

@ekinesme and all the authors, thank you for the excellent work. Congratulations!
