### 2026-04-24 #10136 @espottesmith
Hey @mbagheri20 @colinbousige @jtv3 @shehan807 @muralidhar-nalabothula

We're looking for an additional reviewer for this submission to JOSS. Can any of y'all contribute your expertise to this manuscript/code?
