### 2026-05-15 c5ee919
make installation on apple silicon simpler
