### 2026-05-15 cb6fe49
add Documentation for mac install
