+++ cmake/greenx_dependency.cmake
@@ -22,6 +22,7 @@ else()
                    -DPAW_COMPONENT=OFF
                    -DCOMPILE_SUBMODULES=OFF
                    -DENABLE_GNU_GMP=OFF
+                   -DENABLE_GREENX_CTEST=OFF
         BUILD_COMMAND make -j 
         INSTALL_COMMAND make install && rm -rf ${CMAKE_BINARY_DIR}/external/greenx/src/greenx/.git/
         GIT_SUBMODULES ""
