### 2026-05-15 edb3a4c
add Documentation for mac install
