### 2026-05-15 b7d3d79
uncomment macOS options
