+++ initial_cache.example.cmake
@@ -31,12 +31,12 @@
 #  MacOS + Homebrew (gfortran + OpenBLAS + fftw)
 ################################################################################
 
-# specifying the fortran compiler
-set(CMAKE_Fortran_COMPILER "gfortran" CACHE FILEPATH "Fortran compiler")
-
-# specifying library paths installed with Homebrew:
-#       $ brew install gcc openblas fftw
-set(FFTW_LIBRARY_DIR "/opt/homebrew/opt/fftw/lib/" CACHE PATH "Path to FFTW library directory")
-set(FFTW_INCLUDE_DIR "/opt/homebrew/opt/fftw/include/" CACHE PATH "Path to FFTW include directory")
-set(BLA_VENDOR "OpenBLAS" CACHE STRING "BLAS vendor")
-set(CMAKE_PREFIX_PATH "/opt/homebrew/opt/openblas" CACHE PATH "Path to OpenBLAS installation")
+#   # specifying the fortran compiler
+#   set(CMAKE_Fortran_COMPILER "gfortran" CACHE FILEPATH "Fortran compiler")
+#   
+#   # specifying library paths installed with Homebrew:
+#   #       $ brew install gcc openblas fftw
+#   set(FFTW_LIBRARY_DIR "/opt/homebrew/opt/fftw/lib/" CACHE PATH "Path to FFTW library directory")
+#   set(FFTW_INCLUDE_DIR "/opt/homebrew/opt/fftw/include/" CACHE PATH "Path to FFTW include directory")
+#   set(BLA_VENDOR "OpenBLAS" CACHE STRING "BLAS vendor")
+#   set(CMAKE_PREFIX_PATH "/opt/homebrew/opt/openblas" CACHE PATH "Path to OpenBLAS installation")
