+++ _sources/file_formats.md.txt
@@ -1,10 +1,22 @@
-# File Formats
+# File Formats and Example Scripts
 
-`vibrant` processes multiple data files, such as the dipole moments, polarizabilities, and forces, in order to compute the desired vibrational spectrum. These files must follow specific formats which we describe in this section.
+`vibrant` processes multiple data files, such as the dipole moments, polarizabilities, and forces, in order to compute the desired vibrational spectrum. In principle, these data can be obtained from any quantum chemistry package, provided that the appended data files follow the input formats expected by `vibrant`, which are described in this section.
 
 More information on all available keywords can be found on [Keyword Glossary](Keyword_Glossary.rst) and all complete example input files are available on [Examples](Examples.md).
 
-## 1) Coordinates of atomic positions (.xyz files)
+## Example scripts
+
+Although the ab initio data can, in principle, be obtained using various quantum chemistry packages, most of our DFT calculations are performed with CP2K or FHI-aims. Therefore, we provide a set of utility Bash scripts that can read selected CP2K and FHI-aims output files and append the required data in the appropriate format for `vibrant`. These scripts are intended both to assist users of these packages and to serve as example starting points for users working with other electronic structure codes. The folder containing the example scripts and data files can be downloaded [here](_static/example_scripts_vibrant.tar.gz).
+
+The folder contains three directories:
+
+1. `static_calculations` → Contains four Bash scripts for static spectrum calculations: one for appending CP2K forces, one for FHI-aims forces, one for CP2K dipole moments, and one for CP2K DFPT polarizabilities.  Additionally, directories containing CP2K and FHI-aims outputs for finite-displaced structures of the water molecule are provided to demonstrate how the data can be correctly appended by looping over these structures. In practice, these directories would also contain the corresponding displaced geometries required to perform the DFT calculations. Such directory structures can either be created manually by the user or generated automatically, for example, using the `get_vibrations.py` utility of [FHI-aims](https://fhi-aims.org/uploads/manual/Ch4/S6.html), prior to the DFT calculations. 
+2. `md_append` → Contains two subdirectories that provide Bash scripts for reading MD-based Berry-phase dipole moments and DFPT polarizability data from CP2K outputs, along with example CP2K output files.
+3. `rt-tddft_dipoles` → Contains a Bash script for reading and appending dipole moments from a CP2K RT-TDDFT output file. The provided script loops over MD frames; however, it can easily be adapted to loop over displaced structures for static calculations, following a similar structure to that described in Step 1 in this list. An example CP2K RT-TDDFT output file is also provided.
+
+## Formatting of the input data files
+
+### 1) Coordinates of atomic positions (.xyz files)
 
 Static spectral calculations that are based on the normal mode analysis require the cartesian coordinates (in Å) of each atom in the system to be provided in a standard `.xyz` [format](https://en.wikipedia.org/wiki/XYZ_file_format).
 
@@ -24,9 +36,9 @@ Similarly, the power spectrum calculations requires that the user provides the a
 
 An example file can be downloaded from [here](_static/force.data).
 
-## 2) Files for normal mode analysis
+### 2) Files for normal mode analysis
 
-### 2.1 Atomic forces
+#### 2.1 Atomic forces
 
[…]
