### 2026-05-18 439c58b
add example bash scripts
