### 2026-05-20 82451a2
export fortran compiler to greenx
