+++ .github/workflows/deploy-website.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-      - documentation 
   pull_request:
     branches:
       - main
