+++ JOSS/paper.md
@@ -64,15 +64,15 @@ Computational simulation of vibrational spectroscopy is often not straightforwar
 
 Post-processing programs that can handle these properties are often limited in scope and scattered across different software packages. In this work, we introduce Vibrant, a computational tool which bridges this gap by performing a series of post-processing procedures on dipole moment and polarizability data to generate MD-based or static vibrational spectra. Our code can provide these spectra for gases, liquids and a wide range of materials. Figure&#160;1 provides an overview of Vibrant's major functionalities.
 
-The first panel (top-left) summarizes Vibrant’s frequency analysis capabilities, which include static calculations within the harmonic approximation and post-processing of MD trajectories. In the static approach, conventional normal mode analysis is used to obtain vibrational frequencies and normal mode coordinates, while in the MD-based approach, normal mode derivatives are replaced by Fourier transforms of time-autocorrelation functions, yielding power spectra [@thomas2013computing; @gordon1965molecular].
+ ![ Overview of Vibrant's functionalities, summarizing: (top-left) its ability to generate both static and MD-based vibrational spectra, (top-right) the corresponding spectral intensities available within Vibrant, (bottom-left) the calculation of subspectra for user-defined fragments, and (bottom-right) its integration with Padé approximants. More information about the theoretical background and calculation procedures is available on the [Vibrant website](https://lstsgroup.github.io/vibrant/index.html).](flowchart.png){label="overview"}
+
+The first panel (top-left) summarizes Vibrant’s frequency analysis capabilities, which include static calculations within the harmonic approximation and post-processing of MD trajectories. In the static approach, conventional normal mode analysis is used to obtain vibrational frequencies and normal mode coordinates from energy derivatives with respect to displaced coordinates. In the MD-based approach, vibrational spectra are instead obtained from Fourier transforms of position time-autocorrelation functions, yielding power spectra [@thomas2013computing; @gordon1965molecular].
 
 The second panel (top-right) highlights Vibrant’s functionality for computing IR, non-resonant Raman and RR spectral intensities, which can be obtained from either static or MD-based calculations. For IR and non-resonant Raman spectroscopy, Vibrant can handle different types of dipole moments and polarizabilities. Dipole moments may be provided as Berry-phase dipoles or generated from Wannier centers, while polarizabilities can be taken from density functional perturbation theory (DFPT) or computed from induced dipole moments via a finite-difference approach. RR intensities are obtained by post-processing time-dependent (Berry-phase) dipole trajectories obtained from RT-TDDFT simulations. Although not shown in the panel, Vibrant also supports the computation of absorption spectra from RT-TDDFT dipole data. 
 
 The third panel (bottom-left) displays Vibrant's ability to dissect MD-based spectra via computing the subspectra for user-defined molecular fragments. This feature facilitates the analysis of the spectra, and is especially useful for characterizing guest-host systems, evaluating solvent contributions and molecular materials [@bas2024robust; @guo2010cyclodextrin; @sun2020covalent].
 
-The last panel (bottom-right) demonstrates the use of Padé approximants in absorption or RR spectra calculations, implemented in Vibrant through integrating the GreenX library’s analytic continuation component [@leucke2025analytic]. In RT-TDDFT simulations, frequency resolution strongly depends on the trajectory length, and shorter trajectories often struggle with poor spectral resolution [@bruner2016accelerated]. Applying Padé interpolation is particularly useful for achieving spectral convergence from short RT-TDDFT trajectories [@bruner2016accelerated; @mattiat2018efficient]. As demonstrated in the bottom-right panel, Padé interpolation enables convergence and improved frequency resolution even for relatively short trajectories, resulting in a fivefold reduction of overall computational cost.
-
[…]
