+++ JOSS/paper.bib
@@ -11,7 +11,7 @@
 }
 @article{brehm2020travis,
   title={TRAVIS—A free analyzer for trajectories from molecular simulation},
-  author={Brehm, M and Thomas, M and Gehrke, S and Kirchner, B},
+  author={Brehm, M. and Thomas, M. and Gehrke, S. and Kirchner, B.},
   journal={The Journal of Chemical Physics},
   volume={152},
   number={16},
@@ -20,7 +20,7 @@
   publisher={AIP Publishing},
   doi={10.1063/5.0005078},
 }
-@misc{brehm2011travis,
+@article{brehm2011travis,
   author = {Brehm, Martin and Kirchner, Barbara},
   title = {TRAVIS - A Free Analyzer and Visualizer for Monte Carlo and Molecular Dynamics Trajectories},
   journal = {Journal of Chemical Information and Modeling},
@@ -116,7 +116,7 @@
   url          = {https://doi.org/10.5281/zenodo.4026342},
 }
 @book{long2002raman,
-  title={The Raman Effect},
+  title = {{The Raman Effect}},
   author={Long, Derek A},
   year={2002},
   publisher={John Wiley \& Sons Ltd},
@@ -290,6 +290,8 @@
   author  = {Reichmayr, F. and Wolf, D. and Zhan, G. and Herzog, M. and Wang, M. and Dong, R. and Lubk, A. and Weidinger, I.},
   title   = {Raman Marker Bands for In-situ Quality Control During Synthesis of Two-Dimensional Conjugated Metal-Organic Frameworks},
   journal = {Advanced Materials Interfaces},
+  volume = {12},
+  number = {24},
   year    = {2025},
   pages   = {e00686},
   doi     = {10.1002/admi.202500686}
