+++ JOSS/paper.bib
@@ -9,7 +9,7 @@
   publisher={ACS Publications},
   doi = {10.1021/acs.jctc.4c01021},
 }
-@article{brehm2020travis,
+@article{brehm2020,
   title={TRAVIS—A free analyzer for trajectories from molecular simulation},
   author={Brehm, M. and Thomas, M. and Gehrke, S. and Kirchner, B.},
   journal={The Journal of Chemical Physics},
@@ -20,14 +20,14 @@
   publisher={AIP Publishing},
   doi={10.1063/5.0005078},
 }
-@article{brehm2011travis,
-  author = {Brehm, Martin and Kirchner, Barbara},
-  title = {TRAVIS - A Free Analyzer and Visualizer for Monte Carlo and Molecular Dynamics Trajectories},
-  journal = {Journal of Chemical Information and Modeling},
-  volume = {51},
-  number = {8},
-  pages = {2007--2023},
-  year = {2011},
+@article{brehm2011,
+  title={TRAVIS - A Free Analyzer and Visualizer for Monte Carlo and Molecular Dynamics Trajectories},
+  author={Brehm, Martin and Kirchner, Barbara},
+  journal={Journal of Chemical Information and Modeling},
+  volume={51},
+  number={8},
+  pages={2007--2023},
+  year={2011},
   doi={10.1021/ci200217w},
 }
 @article{thomas2013computing,
