+++ JOSS/paper.bib
@@ -9,7 +9,7 @@
   publisher={ACS Publications},
   doi = {10.1021/acs.jctc.4c01021},
 }
-@article{brehm2020,
+@article{brehm2020travis,
   title={TRAVIS—A free analyzer for trajectories from molecular simulation},
   author={Brehm, M. and Thomas, M. and Gehrke, S. and Kirchner, B.},
   journal={The Journal of Chemical Physics},
@@ -20,9 +20,9 @@
   publisher={AIP Publishing},
   doi={10.1063/5.0005078},
 }
-@article{brehm2011,
+@article{brehm2011travis,
   title={TRAVIS - A Free Analyzer and Visualizer for Monte Carlo and Molecular Dynamics Trajectories},
-  author={Brehm, Martin and Kirchner, Barbara},
+  author={Brehm, M. and Kirchner, B.},
   journal={Journal of Chemical Information and Modeling},
   volume={51},
   number={8},
