### 2026-05-28 c3dd4bc
figure positioning
