+++ JOSS/paper.md
@@ -64,7 +64,7 @@ Computational simulation of vibrational spectroscopy is often not straightforwar
 
 Post-processing programs that can handle these properties are often limited in scope and scattered across different software packages. In this work, we introduce Vibrant, a computational tool which bridges this gap by performing a series of post-processing procedures on dipole moment and polarizability data to generate MD-based or static vibrational spectra. Our code can provide these spectra for gases, liquids and a wide range of materials. Figure&#160;1 provides an overview of Vibrant's major functionalities.
 
- ![ Overview of Vibrant's functionalities, summarizing: (top-left) its ability to generate both static and MD-based vibrational spectra, (top-right) the corresponding spectral intensities available within Vibrant, (bottom-left) the calculation of subspectra for user-defined fragments, and (bottom-right) its integration with Padé approximants. More information about the theoretical background and calculation procedures is available on the [Vibrant website](https://lstsgroup.github.io/vibrant/index.html).](flowchart.png){label="overview"}
+ ![ Overview of Vibrant's functionalities, summarizing: (top-left) its ability to generate both static and MD-based vibrational spectra, (top-right) the corresponding spectral intensities available within Vibrant, (bottom-left) the calculation of subspectra for user-defined fragments, and (bottom-right) its integration with Padé approximants. More information about the theoretical background and calculation procedures is available on the [Vibrant website](https://lstsgroup.github.io/vibrant/index.html).](flowchart.png){#overview fig-pos="t"}
 
 The first panel (top-left) summarizes Vibrant’s frequency analysis capabilities, which include static calculations within the harmonic approximation and post-processing of MD trajectories. In the static approach, conventional normal mode analysis is used to obtain vibrational frequencies and normal mode coordinates from energy derivatives with respect to displaced coordinates. In the MD-based approach, vibrational spectra are instead obtained from Fourier transforms of position time-autocorrelation functions, yielding power spectra [@thomas2013computing; @gordon1965molecular].
 
