+++ JOSS/paper.md
@@ -62,9 +62,9 @@ Vibrational spectroscopy, particularly IR and Raman techniques, provides valuabl
 
 Computational simulation of vibrational spectroscopy is often not straightforward. Most quantum chemistry packages that perform electronic structure calculations provide only the dipole moments or polarizability tensors, rather than the final spectrum itself. The challenge becomes greater when one goes beyond the static regime to compute vibrational spectra, such as when performing MD simulations to capture anharmonic modes or solvent effects, or real-time time-dependent density functional theory (RT-TDDFT) simulations to access excited state dynamics, relevant for RR spectroscopy.
 
-Post-processing programs that can handle these properties are often limited in scope and scattered across different software packages. In this work, we introduce Vibrant, a computational tool which bridges this gap by performing a series of post-processing procedures on dipole moment and polarizability data to generate MD-based or static vibrational spectra. Our code can provide these spectra for gases, liquids and a wide range of materials. Figure&#160;1 provides an overview of Vibrant's major functionalities.
-
+Post-processing programs that can handle these properties are often limited in scope and scattered across different software packages. In this work, we introduce Vibrant, a computational tool which bridges this gap by performing a series of post-processing procedures on dipole moment and polarizability data to generate MD-based or static vibrational spectra. Our code can provide these spectra for gases, liquids and a wide range of materials. Figure&#160;1 provides 
  ![ Overview of Vibrant's functionalities, summarizing: (top-left) its ability to generate both static and MD-based vibrational spectra, (top-right) the corresponding spectral intensities available within Vibrant, (bottom-left) the calculation of subspectra for user-defined fragments, and (bottom-right) its integration with Padé approximants. More information about the theoretical background and calculation procedures is available on the [Vibrant website](https://lstsgroup.github.io/vibrant/index.html).](flowchart.png){#overview fig-pos="t"}
+ an overview of Vibrant's major functionalities.
 
 The first panel (top-left) summarizes Vibrant’s frequency analysis capabilities, which include static calculations within the harmonic approximation and post-processing of MD trajectories. In the static approach, conventional normal mode analysis is used to obtain vibrational frequencies and normal mode coordinates from energy derivatives with respect to displaced coordinates. In the MD-based approach, vibrational spectra are instead obtained from Fourier transforms of position time-autocorrelation functions, yielding power spectra [@thomas2013computing; @gordon1965molecular].
 
