### 2026-06-04 3b83979
add the JOSS citation to the documentation website
