+++ docs/Cite.md
@@ -1,15 +1,17 @@
 # Cite 
 
-If you use the code, please cite the following reference: 
+Our paper on `vibrant`, "Vibrant: A Post-Processing Tool for Computational Vibrational Spectroscopy of Molecules, Liquids and Materials" was published in _Journal of Open Source Software_. You can access the paper [here](https://joss.theoj.org/papers/10.21105/joss.10136).  
+
+If you use `vibrant` in your research, please cite the following reference: 
 
 ```bash 
-@article{winogradow2025vibrant,
+@article{winogradow2026vibrant,
   title={Vibrant: A Post-Processing Tool for Computational Vibrational Spectroscopy of Molecules, Liquids and Materials},
   author={Ekin Winogradow, Johannes Scheffler, Moritz Leucke, Thomas Heine, and Dorothea Golze},
-  journal={Journal of open source software},
-  volume={_},
-  number={__},
-  pages={__},
-  year={2025}
+  journal={Journal of Open Source Software},
+  volume={11},
+  number={122},
+  pages={10136},
+  year={2026}
 }
 ```
