+++ .github/workflows/deploy-website.yml
@@ -4,6 +4,7 @@ on:
   push:
     branches:
       - main
+      - update-doc 
   pull_request:
     branches:
       - main
