+++ _sources/index.rst.txt
@@ -9,7 +9,7 @@
 
 Overview 
 ===========
-The ``vibrant`` code is a post-processing tool for calculating vibrational spectra from ab initio data.
+The ``vibrant`` code is a post-processing tool for calculating vibrational spectra from atomistic simulation data.
 
 .. raw:: html
 
@@ -20,12 +20,12 @@ The ``vibrant`` code is a post-processing tool for calculating vibrational spect
    </div>
 Features
 ===========
-   * **IR and Raman spectra** from static calculations or *ab initio* molecular dynamics (MD).  
+   * **Vibrational frequencies, IR and Raman spectra** from static calculations or molecular dynamics (MD) simulations.  
    * **Resonance Raman spectra** via efficient Padé approximation or explicit frequency sampling.  
-   * **Absorption spectra** computed alongside vibrational data for comprehensive analysis. 
+   * Calculation of **Absorption spectra** alongside vibrational data for comprehensive analysis. 
    * **Dissection of MD-based spectra** into user-defined molecular fragments.
    * **Fast and efficient** post-processing of large systems and long trajectories.
-   * Applicable to **single molecules, liquids and materials**
+   * Applicable to **single molecules, liquids and materials**.
 
 Compatibility and Applications
 ===========
