### 2026-06-04 b536756
add the JOSS citation to the documentation website
