### 2026-06-04 74b7c72
add the JOSS citation to the documentation website
