+++ .github/workflows/deploy-website.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-      - update-doc 
   pull_request:
     branches:
       - main
